+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Update fill light and bi-color light icons
 - Set dark mode as default to prevent bright flash on startup
 - Updated scene placeholder dashed border to be more symmetrical
+- Fade out red error outline on faulty inputs
 
 ### Fixes